Add files via upload
I made a one of popular game " Simon " that tests your short-term memory skills. It has four coloured buttons that light up and make sounds in a random sequence. You have to repeat the sequence by pressing the buttons in the same order. The game gets harder as the sequence gets longer and faster. If our sequence patter was wrong then the game over otherwise the sequences are generated infinitely.